frontend/core/blocks/flip-image-horizontally-20250115-142546-791/requirements.txt
numpy==1.26.4
Pillow==11.1.0